+++ etc/launch-with-python-udf-server/python-udf-server.toml
@@ -2,4 +2,4 @@ service-type = "PYTHON_UDF"
 
 [python-udf-server]
 address = "127.0.0.1:50051"
-path = "pkg/pythonserver/python"
+path = "pkg/udf/pythonservice/pyserver"
