+++ pkg/cnservice/server.go
@@ -48,6 +48,8 @@ import (
 	"github.com/matrixorigin/matrixone/pkg/txn/client"
 	"github.com/matrixorigin/matrixone/pkg/txn/rpc"
 	"github.com/matrixorigin/matrixone/pkg/txn/storage/memorystorage"
+	"github.com/matrixorigin/matrixone/pkg/udf"
+	"github.com/matrixorigin/matrixone/pkg/udf/pythonservice"
 	"github.com/matrixorigin/matrixone/pkg/util/address"
 	"github.com/matrixorigin/matrixone/pkg/util/executor"
 	"github.com/matrixorigin/matrixone/pkg/util/profile"
@@ -139,10 +141,22 @@ func NewService(
 		MaxSize:        pu.SV.AutoIncrCacheSize,
 	}
 
+	// init UdfService
+	srv.udfService = udf.NewService()
+	// add python client to handle python udf
+	if srv.cfg.PythonUdfClient.ServerAddress != "" {
+		pc, err := pythonservice.NewClient(srv.cfg.PythonUdfClient)
+		if err != nil {
+			panic(err)
+		}
+		srv.udfService.AddPythonUdfClient(pc)
+	}
+
 	srv.pu = pu
 	srv.pu.LockService = srv.lockService
 	srv.pu.HAKeeperClient = srv._hakeeperClient
 	srv.pu.QueryService = srv.queryService
+	srv.pu.UdfService = srv.udfService
 	srv._txnClient = pu.TxnClient
 
 	if err = srv.initMOServer(ctx, pu, srv.aicm); err != nil {
@@ -179,6 +193,7 @@ func NewService(
 		lockService lockservice.LockService,
 		queryService queryservice.QueryService,
 		hakeeper logservice.CNHAKeeperClient,
+		udfService udf.Service,
 		cli client.TxnClient,
 		aicm *defines.AutoIncrCacheManager,
 		messageAcquirer func() morpc.Message) error {
@@ -360,6 +375,7 @@ func (s *service) handleRequest(
 			s.lockService,
 			s.queryService,
 			s._hakeeperClient,
+			s.udfService,
 			s._txnClient,
 			s.aicm,
 			s.acquireMessage)
@@ -661,6 +677,7 @@ func (s *service) initInternalSQlExecutor(mp *mpool.MPool) {
 		s.fileService,
 		s.queryService,
 		s._hakeeperClient,
+		s.udfService,
 		s.aicm)
 	runtime.ProcessLevelRuntime().SetGlobalVariables(runtime.InternalSQLExecutor, exec)
 }
