+++ pkg/config/configuration.go
@@ -27,6 +27,7 @@ import (
 	"github.com/matrixorigin/matrixone/pkg/queryservice"
 	"github.com/matrixorigin/matrixone/pkg/taskservice"
 	"github.com/matrixorigin/matrixone/pkg/txn/client"
+	"github.com/matrixorigin/matrixone/pkg/udf"
 	"github.com/matrixorigin/matrixone/pkg/util/toml"
 	"github.com/matrixorigin/matrixone/pkg/vm/engine"
 )
@@ -667,6 +668,8 @@ type ParameterUnit struct {
 
 	// QueryService instance
 	QueryService queryservice.QueryService
+
+	UdfService udf.Service
 
 	// HAKeeper client, which is used to get connection ID
 	// from HAKeeper currently.
