+++ pkg/vm/process/types.go
@@ -36,6 +36,7 @@ import (
 	"github.com/matrixorigin/matrixone/pkg/logservice"
 	"github.com/matrixorigin/matrixone/pkg/queryservice"
 	"github.com/matrixorigin/matrixone/pkg/txn/client"
+	"github.com/matrixorigin/matrixone/pkg/udf"
 	"github.com/matrixorigin/matrixone/pkg/vm/engine"
 )
 
@@ -324,6 +325,8 @@ type Process struct {
 	QueryService queryservice.QueryService
 
 	Hakeeper logservice.CNHAKeeperClient
+
+	UdfService udf.Service
 }
 
 type vectorPool struct {
